[core] Add a fuzzing regression test github action (#674)
While fuzzing subzero core before the nanopb 0.4.5 upgrade,
the fuzzer found a bunch of inputs which triggered UBSAN.

This PR creates a regression test that runs the fuzzer with these inputs
and verifies that none of them fail now that we're on nanopb 0.4.5.

If more bad inputs are discovered later, they can be added to this directory
after the corresponding bugs are fixed.

+++ scripts/build_core.sh
@@ -8,9 +8,19 @@ cd core/build
 # For mainnet. Skip compile_db generation
 TARGET=dev CURRENCY=btc-mainnet cmake ../ "$@"
 make
-mv subzero subzero-mainnet
+if [ -f subzero ]; then
+  mv subzero subzero-mainnet
+fi
+if [ -f subzero_fuzzer ]; then
+  mv subzero_fuzzer subzero_fuzzer-mainnet
+fi
 # For testnet. Generate compile_db for clang static analyzer
 make clean
 TARGET=dev CURRENCY=btc-testnet cmake ../ -DCMAKE_EXPORT_COMPILE_COMMANDS=ON "$@"
 make
-mv subzero subzero-testnet
+if [ -f subzero ]; then
+  mv subzero subzero-testnet
+fi
+if [ -f subzero_fuzzer ]; then
+  mv subzero_fuzzer subzero_fuzzer-testnet
+fi
